#8.1

+++ controllers/videoController.js
@@ -32,7 +32,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      res.render('editVideo', { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -75,7 +79,7 @@ export const videoDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await await Video.findById(id).populate('creator');
+    const video = await Video.findById(id).populate('creator');
     console.log(video);
     res.render('videoDetail', { pageTitle: video.title, video });
   } catch (error) {
@@ -87,8 +91,14 @@ export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
+  console.log('Delete');
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator.toString() !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
